PluginMenuSlug: move sniff from VIP category to Security category
The original sniff class still exists, but extends the renamed sniff now.

Note: if the sniff was explicitly **excluded** via a custom ruleset, no deprecation warning will be thrown and the exclusion will have no effect. The custom ruleset will need to be updated to use the new sniff name for the exclusion to be reinstated.
