+++ WordPress/Sniffs/VIP/PluginMenuSlugSniff.php
@@ -21,67 +21,45 @@ use WordPress\AbstractFunctionParameterSniff;
  * @since   0.3.0
  * @since   0.11.0 Refactored to extend the new WordPress_AbstractFunctionParameterSniff.
  * @since   0.13.0 Class name changed: this class is now namespaced.
+ *
+ * @deprecated 0.15.0 This sniff has been moved to the `Security` category.
+ *                    This file remains for now to prevent BC breaks.
  */
-class PluginMenuSlugSniff extends AbstractFunctionParameterSniff {
+class PluginMenuSlugSniff extends \WordPress\Sniffs\Security\PluginMenuSlugSniff {
 
 	/**
-	 * The group name for this group of functions.
+	 * Keep track of whether the warning has been thrown to prevent
+	 * the message being thrown for every token triggering the sniff.
 	 *
-	 * @since 0.11.0
+	 * @since 0.15.0
 	 *
-	 * @var string
+	 * @var array
 	 */
-	protected $group_name = 'add_menu_functions';
-
-	/**
-	 * Functions which can be used to add pages to the WP Admin menu.
-	 *
-	 * @since 0.3.0
-	 * @since 0.11.0 Renamed from $add_menu_functions to $target_functions
-	 *               and changed visibility to protected.
-	 *
-	 * @var array <string function name> => <array target parameter positions>
-	 */
-	protected $target_functions = array(
-		'add_menu_page'       => array( 4 ),
-		'add_object_page'     => array( 4 ),
-		'add_utility_page'    => array( 4 ),
-		'add_submenu_page'    => array( 1, 5 ),
-		'add_dashboard_page'  => array( 4 ),
-		'add_posts_page'      => array( 4 ),
-		'add_media_page'      => array( 4 ),
-		'add_links_page'      => array( 4 ),
-		'add_pages_page'      => array( 4 ),
-		'add_comments_page'   => array( 4 ),
-		'add_theme_page'      => array( 4 ),
-		'add_plugins_page'    => array( 4 ),
-		'add_users_page'      => array( 4 ),
-		'add_management_page' => array( 4 ),
-		'add_options_page'    => array( 4 ),
+	private $thrown = array(
+		'DeprecatedSniff' => false,
 	);
 
 	/**
-	 * Process the parameters of a matched function.
+	 * Don't use.
 	 *
-	 * @since 0.11.0
+	 * @deprecated 0.15.0
 	 *
-	 * @param int    $stackPtr        The position of the current token in the stack.
-	 * @param array  $group_name      The name of the group which was matched.
-	 * @param string $matched_content The token content (function name) which was matched.
-	 * @param array  $parameters      Array with information about the parameters.
+	 * @param int $stackPtr The position of the current token in the stack.
 	 *
-	 * @return void
+	 * @return void|int
 	 */
-	public function process_parameters( $stackPtr, $group_name, $matched_content, $parameters ) {
-		foreach ( $this->target_functions[ $matched_content ] as $position ) {
-			if ( isset( $parameters[ $position ] ) ) {
-				$file_constant = $this->phpcsFile->findNext( T_FILE, $parameters[ $position ]['start'], ( $parameters[ $position ]['end'] + 1 ) );
+	public function process_token( $stackPtr ) {
+		if ( false === $this->thrown['DeprecatedSniff'] ) {
+			$this->phpcsFile->addWarning(
+				'The "WordPress.VIP.PluginMenuSlug" sniff has been renamed to "WordPress.Security.PluginMenuSlug". Please update your custom ruleset.',
+				0,
+				'DeprecatedSniff'
+			);
 
-				if ( false !== $file_constant ) {
-					$this->phpcsFile->addError( 'Using __FILE__ for menu slugs risks exposing filesystem structure.', $stackPtr, 'Using__FILE__' );
-				}
-			}
+			$this->thrown['DeprecatedSniff'] = true;
 		}
+
+		return parent::process_token( $stackPtr );
 	}
 
 } // End class.
