+++ WordPress/Tests/Security/PluginMenuSlugUnitTest.php
@@ -7,7 +7,7 @@
  * @license https://opensource.org/licenses/MIT MIT
  */
 
-namespace WordPress\Tests\VIP;
+namespace WordPress\Tests\Security;
 
 use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
 
@@ -18,8 +18,7 @@ use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
  *
  * @since   0.3.0
  * @since   0.13.0 Class name changed: this class is now namespaced.
- * @since   0.15.0 The sniff has been deprecated. This unit test file now
- *                 only tests that the deprecation warnings are correctly thrown.
+ * @since   0.15.0 This sniff has been moved from the `VIP` category to the `Security` category.
  */
 class PluginMenuSlugUnitTest extends AbstractSniffUnitTest {
 
@@ -29,7 +28,12 @@ class PluginMenuSlugUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of errors>
 	 */
 	public function getErrorList() {
-		return array();
+		return array(
+			3 => 1,
+			5 => 1,
+			9 => 2,
+		);
+
 	}
 
 	/**
@@ -38,9 +42,7 @@ class PluginMenuSlugUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of warnings>
 	 */
 	public function getWarningList() {
-		return array(
-			1 => 1,
-		);
+		return array();
 
 	}
