+++ WordPress/ruleset.xml
@@ -110,4 +110,14 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.VIP.PluginMenuSlug.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.VIP.PluginMenuSlug.Using__FILE__">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
